:hammer: Update bookCatalog/table component

+++ client/components/bookCatalog/table.jsx
@@ -1,5 +1,8 @@
 import React from 'react';
+import moment from 'moment';
+
 import style from '../../styles/components/bookCatalog/table.css';
+import handleGetLoan from '../../helpers/handleGetLoan';
 
 function Table({
   catalogs,
@@ -8,23 +11,19 @@ function Table({
   setUpdateData,
   setUpdateBookIsOpen,
   setDetails,
+  page,
+  setConfirmDelete,
 }) {
   const isDev = process.env.NODE_ENV === 'development';
-
-  const formatDate = (args) => {
-    const date = new Date(args).toLocaleDateString([], {
-      day: '2-digit',
-      month: 'short',
-      year: 'numeric',
-    });
-
-    return date;
-  }
+  const pagLimit = JSON.parse(localStorage.getItem('pag'));
 
   const handleDeleteBook = async (args) => {
     try {
       const token = localStorage.getItem('token');
       const url = isDev ? 'http://localhost:8000/api/books' : '/api/books';
+
+      const loan = await handleGetLoan(args.bookCode);
+      if (loan.length > 0) throw loan;
 
       const request = await (await fetch(url, {
         method: 'delete',
@@ -43,16 +42,22 @@ function Table({
       ctx.style = 'opacity: 0';
 
       setTimeout(() => {
-        handleGetBookCatalogs();
+        handleGetBookCatalogs(pagLimit.book);
         handleInfoboxData();
 
         setTimeout(() => {
           ctx.style = 'opacity: 1';
         }, 500);
-      }, 1000);
+      }, 500);
     }
     catch (error0) {
-      console.error(error0.message);
+      setConfirmDelete(() => ({
+        data: {
+          index: args.index + 1,
+          bookCode: args.bookCode,
+        },
+        isOpen: true,
+      }))
     }
   }
 
@@ -87,13 +92,17 @@ function Table({
         {
           catalogs.map((item, index) => (
             <tr className={style.row} key={item.bookCode}>
-              <td className={style.column} onClick={() => handleOpenDetails(item)} aria-hidden="true">{index + 1}</td>
+              <td className={style.column} onClick={() => handleOpenDetails(item)} aria-hidden="true">
+                {page > 1 ? (pagLimit.book * (page - 1)) + (index + 1) : index + 1}
+              </td>
               <td className={style.column} onClick={() => handleOpenDetails(item)} aria-hidden="true">{item.bookCode}</td>
               <td className={style.column} onClick={() => handleOpenDetails(item)} aria-hidden="true">{item.title}</td>
               <td className={style.column} onClick={() => handleOpenDetails(item)} aria-hidden="true">{item.author}</td>
-              <td className={style.column} onClick={() => handleOpenDetails(item)} aria-hidden="true">{item.publisher}</td>
+              <td className={style.column} onClick={() => handleOpenDetails(item)} aria-hidden="true">
+                {item.publisher ? item.publisher : '-'}
+              </td>
               <td className={style.column} onClick={() => handleOpenDetails(item)} aria-hidden="true">{item.stock}</td>
-              <td className={style.column} onClick={() => handleOpenDetails(item)} aria-hidden="true">{formatDate(item.publicationDate)}</td>
+              <td className={style.column} onClick={() => handleOpenDetails(item)} aria-hidden="true">{moment(item.publicationDate).format('ll')}</td>
               <td className={`${style.column} ${style.action}`}>
                 <button
                   type="button"
